eb_assessee_gym/pyproject.toml
[project]
name = "eb-assessee-gym"
version = "0.1.0"
description = "gym-style a2a agent implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.22",
    "mcp>=1.26.0",
    "openenv-core>=0.2.1",
    "openenv-echo-env",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
openenv-echo-env = { git = "https://huggingface.co/spaces/openenv/echo_env" }
